Return from onTouchEnded if highlighted is not true.
This allows us to set highlighted back to false in code and have this stop a button press. Useful for emulating the behaviour of the iOS home screen. This is: press a button - it's shown as down, release to trigger it, or swipe sideways to swipe between screens. The swipe is used to cancel the button down.